수정

+++ Form1.cs
@@ -23,37 +23,47 @@ namespace Satisfactory_서버용
         private void button1_Click(object sender, EventArgs e)
         {
             MessageBox.Show("잠깐 steamcmd.exe 가 있어야 됩니다.");
+            B:
+            Process[] u = Process.GetProcessesByName("steamcmd");
+           if (u.GetLength(0) > 0)
+            {
+                MessageBox.Show("어라 steamcmd.exe 가 실행 되고 있습니다 \r 강재 종료 됩니다.!");
+                u[0].CloseMainWindow();                
+                goto B;
+            }
+            else { 
             AC:
             OpenFileDialog A = new OpenFileDialog();
             A.Title = "steamcmd.exe 선택하세요";
             A.FileName = "steamcmd.exe";
             A.Filter = "실행파일(*.exe)|*.exe;";
             DialogResult d = A.ShowDialog();
-            if (d == DialogResult.OK)
-            {
-                if (Path.GetFileName(A.FileName) == "steamcmd.exe")
+                if (d == DialogResult.OK)
                 {
-                    string fileName = A.FileName;
-                    string[] lines = { "@echo off", "steamcmd.exe +login anonymous +force_install_dir C:\\Satisfactory_Dedicated +app_update 1690800 +quit", "pause" };
-                    File.WriteAllLines(Path.GetDirectoryName(A.FileName) + "\\commd.bat", lines);
-                    Process p = new Process();
-                    p.StartInfo.FileName = "commd.bat";
-                    p.StartInfo.WorkingDirectory = Path.GetDirectoryName(fileName);
-                    p.Start();
-                    p.WaitForExit(1000);
-                    File.Delete(Path.GetDirectoryName(fileName) + "\\commd.bat");
+                    if (Path.GetFileName(A.FileName) == "steamcmd.exe")
+                    {
+                        string fileName = A.FileName;
+                        string[] lines = { "@echo off", "steamcmd.exe +login anonymous +force_install_dir C:\\Satisfactory_Dedicated +app_update 1690800 +quit", "pause" };
+                        File.WriteAllLines(Path.GetDirectoryName(A.FileName) + "\\commd.bat", lines);
+                        Process p = new Process();
+                        p.StartInfo.FileName = "commd.bat";
+                        p.StartInfo.WorkingDirectory = Path.GetDirectoryName(fileName);
+                        p.Start();
+                        p.WaitForExit(1000);
+                        File.Delete(Path.GetDirectoryName(fileName) + "\\commd.bat");
+                    }
+                    else
+                    {
+                        MessageBox.Show("앗 stamcmd.exe를 선택을 안하셨습니다.");
+                        goto AC;
+                    }
                 }
-                else
+                else if (d == DialogResult.Cancel)
                 {
-                    MessageBox.Show("앗 stamcmd.exe를 선택을 안하셨습니다.");
-                    goto AC;                    
-                }  
-            }
-            else if (d == DialogResult.Cancel)
-            {
-                MessageBox.Show("파일이 없습니까?");
-                Form3 m = new Form3();
-                m.ShowDialog();
+                    MessageBox.Show("파일이 없습니까?");
+                    Form3 m = new Form3();
+                    m.ShowDialog();
+                }
             }
         }
